Refactor Ollama integration and enhance logging for model handling
- Updated import paths for `pullModel` and added error handling during model pulling.
- Enhanced logging to provide better insights when starting the Ollama server and pulling models.
- Simplified the `runOllama` function by consolidating model verification and process creation logic.
- Improved error handling for server startup and model loading processes to facilitate easier debugging.
