+++ Frontend/src/electron/ollama/ollamaPath.ts
@@ -1,0 +1,4 @@
+import { platform } from "os";
+
+export const getOllamaPath = () =>
+  platform() === "darwin" ? "/usr/local/bin/ollama" : "ollama";
